Fixed up some problems with liking and liked content stream.

+++ victorious/victorious/Operations/Request Operations/RemoveStreamItemsOperation.swift
@@ -9,27 +9,31 @@
 import Foundation
 import VictoriousIOSSDK
 
-/// Remotes a stream item from the stream and deletes it
 class RemoveStreamItemOperation: Operation {
     
     private let streamItemIDs: [String]
     
     let persistentStore: PersistentStoreType = PersistentStoreSelector.defaultPersistentStore
     
-    init( streamItemIDs: [String]) {
-        self.streamItemIDs = streamItemIDs
+    convenience init(streamItemID: String) {
+        self.init(streamItemIDs: [streamItemID])
     }
     
-    convenience init( streamItemID: String) {
-        self.init( streamItemIDs: [streamItemID] )
+    required init(streamItemIDs: NSArray) {
+        self.streamItemIDs = streamItemIDs as? [String] ?? []
     }
     
     override func main() {
         persistentStore.createBackgroundContext().v_performBlockAndWait() { context in
+            
+            if self.streamItemIDs.isEmpty {
+                return
+            }
+            
             for streamItemID in self.streamItemIDs {
                 if let streamItem: VStreamItem = context.v_findObjects([ "remoteId" : streamItemID ]).first,
                     let streamID = streamItem.streamId,
-                    let stream: VStream = context.v_findObjects([ "remoteId" : streamID ]).first {
+                     let stream: VStream = context.v_findObjects([ "remoteId" : streamID ]).first {
                         stream.v_removeObject( streamItem, from: "streamItems" )
                         context.deleteObject( streamItem )
                 }
